refactors code into new astar file, customEntities file, and the heuristics file
