+++ Bomberman/group25/scenario2_AStarCharacterWithBomb.py
@@ -5,85 +5,8 @@ sys.path.insert(0, '../bomberman')
 # Import necessary stuff
 from entity import CharacterEntity
 import math
-
-
-# this version of monster is used to determine the type of monster when it is supposed to be hidden (see readme)
-class Monster:
-    def __init__(self, x, y):
-        self.prevX = x
-        self.prevY = y
-        self.x = x
-        self.y = y
-        self.velocityX = None
-        self.velocityY = None
-
-        self.timesIActedSmart = 0
-        self.type = None
-        self.path = []
-
-    def getPath(self, wrld):
-        self.checkVelocity(wrld)
-        path = [Node(self.x, self.y)]
-        isOutOfBounds = False
-        i = 1
-        while not isOutOfBounds:
-            (nx, ny) = (self.prevX + self.velocityX*i, self.prevY + self.velocityY*i)
-            # If next pos is out of bounds, must change direction
-            if ((nx < 0) or (nx >= wrld.width()) or
-                    (ny < 0) or (ny >= wrld.height())):
-                isOutOfBounds = True
-            else:
-                path.append(Node(self.x + self.velocityX*i, self.y + self.velocityY*i))
-                i += 1
-        return path
-
-    def checkVelocity(self, wrld):
-        currVelocityX = self.x - self.prevX
-        currVelocityY = self.y - self.prevY
-        if self.velocityX is None or self.must_change_direction(wrld):
-            self.velocityX = currVelocityX
-            self.velocityY = currVelocityY
-            self.prevX = self.x
-            self.prevY = self.y
-
-        elif self.velocityX == currVelocityX and self.velocityY == currVelocityY and self.type != "stupid":
-            self.timesIActedSmart += 1
-            self.prevX = self.x
-            self.prevY = self.y
-
-            # if moved in a manner consistent to self-preserving monster, then it's smart
-            if self.timesIActedSmart > 2:
-                self.type = "smart"
-        elif self.timesIActedSmart < 8:
-            self.type = "stupid"
-        return
-
-    # stolen from Self Preserving Monster
-    def must_change_direction(self, wrld):
-        # Get next desired position
-        (nx, ny) = (self.prevX + self.velocityX, self.prevY + self.velocityY)
-        # If next pos is out of bounds, must change direction
-        if ((nx < 0) or (nx >= wrld.width()) or
-                (ny < 0) or (ny >= wrld.height())):
-            return True
-        # If these cells are an explosion, a wall, or a monster, go away
-        # return (wrld.explosion_at(self.x, self.y) or
-        return (wrld.wall_at(nx, ny) or
-                wrld.exit_at(nx, ny))
-
-
-class Node():
-    def __init__(self, x, y, hval=0, gval=0, parent=None):
-        self.x = x
-        self.y = y
-        self.hval = hval
-        self.gval = gval
-        self.fval = hval + gval
-        self.parent = parent
-
-    # def __eq__(self, other):
-    #    return self.position == other.position
-
+import customEntities
+import astar
 
 class TestCharacter(CharacterEntity):
     def __init__(self, name, avatar, x, y, shouldPanic, distanceStupid, distanceSmart):
@@ -105,8 +28,8 @@ class TestCharacter(CharacterEntity):
     def do(self, wrld):
         ispanicking = False
         # recalculate the AStar path and distance from exit every time
-        self.distanceFromExit = self.distanceBetweenNodes(Node(self.x, self.y), Node(7,18), False)
-        self.calculateCharacterPath(Node(7, 18), wrld, True)
+        self.distanceFromExit = astar.distanceBetweenNodes(customEntities.Node(self.x, self.y), customEntities.Node(7,18), self.monsters, False)
+        self.calculateCharacterPath(customEntities.Node(7, 18), wrld, True)
         self.pathIterator = 0
 
         if self.explosionTimer > 0:
@@ -128,7 +51,7 @@ class TestCharacter(CharacterEntity):
             self.monsters = []
             for key, monsterlist in wrld.monsters.items():
                 for monster in monsterlist:
-                    self.monsters.append(Monster(monster.x, monster.y))
+                    self.monsters.append(customEntities.Monster(monster.x, monster.y))
         else:
             i = 0
             for key, monsterlist in wrld.monsters.items():
@@ -146,7 +69,7 @@ class TestCharacter(CharacterEntity):
 
             # if my path ends at the exit
             selfIsCloserToExitThanMonster = True
-            myPathToExit = self.calculateAStarPath(self, Node(7, 18), wrld, False)
+            myPathToExit = astar.calculateAStarPath(self, customEntities.Node(7, 18), wrld, self.monsters, False)
 
             # if there isn't currently a path to the exit
             if not myPathToExit[0]:
@@ -155,7 +78,7 @@ class TestCharacter(CharacterEntity):
             else:
                 for key, monsterlist in wrld.monsters.items():
                     for monster in monsterlist:
-                        if len(myPathToExit[1]) >= len(self.calculateAStarPath(monster, Node(7, 18), wrld, False)[1]):
+                        if len(myPathToExit[1]) >= len(astar.calculateAStarPath(monster, customEntities.Node(7, 18), wrld, self.monsters, False)[1]):
                             selfIsCloserToExitThanMonster = False
                             break
             if self.shouldPanic and self.bombTimer == 0 and self.x == 0:
@@ -167,7 +90,7 @@ class TestCharacter(CharacterEntity):
                 self.runAway(wrld)
 
             else:
-                self.calculateCharacterPath(Node(7, 18), wrld, False)
+                self.calculateCharacterPath(customEntities.Node(7, 18), wrld, False)
 
         if self.placeBombAtEnd:
             self.place_bomb()
@@ -184,7 +107,7 @@ class TestCharacter(CharacterEntity):
     def checkIfNearMonster(self, node, wrld):
         for monster in self.monsters:
             # if SelfPreservingMonster (smart) monster is within <maxDistance> steps, return true
-            pathBetweenSelfAndMonster = self.calculateAStarPath(node, monster, wrld, False)
+            pathBetweenSelfAndMonster = astar.calculateAStarPath(node, monster, wrld, self.monsters, False)
 
             # if there is path between myself and the monster
             if pathBetweenSelfAndMonster[0]:
@@ -203,7 +126,7 @@ class TestCharacter(CharacterEntity):
     # TODO Implement alpha beta search to try to run away faster
     def runAway(self, wrld):
         # put current position into the path
-        path = [Node(self.x, self.y)]
+        path = [customEntities.Node(self.x, self.y)]
         possibleNodes = []
         shouldRun = False
 
@@ -211,8 +134,8 @@ class TestCharacter(CharacterEntity):
         highestSum = -math.inf
 
         # get the set of neighbors, including ourself
-        neighbors = self.getNeighbors(self, Node(7, 18), wrld)
-        neighbors.append(Node(self.x, self.y))
+        neighbors = astar.getNeighbors(self, customEntities.Node(7, 18), wrld)
+        neighbors.append(customEntities.Node(self.x, self.y))
 
         # iterate over the available spots of the eight cardinal directions
         for node in neighbors:
@@ -244,8 +167,8 @@ class TestCharacter(CharacterEntity):
             node.hval = math.inf
             # for monster in monsterlist:
             for monster in self.monsters:
-                pathBetweenSelfAndMonster = self.calculateAStarPath(node, monster, wrld, False)
-                myDistance = len(self.calculateAStarPath(self, monster, wrld, False)[1])
+                pathBetweenSelfAndMonster = astar.calculateAStarPath(node, monster, wrld, self.monsters, False)
+                myDistance = len(astar.calculateAStarPath(self, monster, wrld, self.monsters, False)[1])
 
 
                 # if there is a path between myself and the monster
@@ -260,7 +183,7 @@ class TestCharacter(CharacterEntity):
                         elif distance < (self.distanceSmart - 2):
                             currSum -= 10
                         currSum += distance
-                        node.hval = self.absoluteDistanceBetweenNodes(node, monster)
+                        node.hval = astar.absoluteDistanceBetweenNodes(node, monster)
                         if len(wrld.bombs.items()) > 0:
                             shouldRun = True
 
@@ -276,11 +199,11 @@ class TestCharacter(CharacterEntity):
                 highestSum = currSum
                 possibleNodes = [node]
 
-        pathToStart = (self.calculateAStarPath(self, Node(0, 4*self.numberOfBombsPlaced), wrld, True))[1]
-        self.calculateCharacterPath(Node(7, 18), wrld, True)
+        pathToStart = (astar.calculateAStarPath(self, customEntities.Node(0, 4*self.numberOfBombsPlaced), wrld, self.monsters, True))[1]
+        self.calculateCharacterPath(customEntities.Node(7, 18), wrld, True)
         if not possibleNodes:
             # accept death.
-            self.path = [Node(self.x, self.y), Node(self.x, self.y)]
+            self.path = [customEntities.Node(self.x, self.y), customEntities.Node(self.x, self.y)]
             return
 
         bestNode = None
@@ -318,8 +241,8 @@ class TestCharacter(CharacterEntity):
 
     # return the character's path to the end
     def calculateCharacterPath(self, endNode, wrld, shouldPathAroundMonsters):
-        startNode = Node(self.x, self.y)
-        path = self.calculateAStarPath(startNode, endNode, wrld, shouldPathAroundMonsters)
+        startNode = customEntities.Node(self.x, self.y)
+        path = astar.calculateAStarPath(startNode, endNode, wrld, self.monsters, shouldPathAroundMonsters)
 
         if path[0]:
             self.path = path[1]
@@ -329,126 +252,14 @@ class TestCharacter(CharacterEntity):
             minNode = startNode
             minNode.hval = math.inf
             for node in path[1]:
-                node.hval = self.absoluteDistanceBetweenNodes(node, endNode)
+                node.hval = astar.absoluteDistanceBetweenNodes(node, endNode)
                 if node.hval < minNode.hval:
                     minNode = node
 
-            if self.absoluteDistanceBetweenNodes(minNode, endNode) == self.absoluteDistanceBetweenNodes(self, endNode) \
+            if astar.absoluteDistanceBetweenNodes(minNode, endNode) == astar.absoluteDistanceBetweenNodes(self, endNode) \
                 and self.bombTimer == 0 and len(wrld.explosions.items()) == 0:
                 self.placeBombAtEnd = True
 
             else:
                 self.calculateCharacterPath(minNode, wrld, True)
 
-
-
-    # take into account walls; if there is a wall in the way, move to it and bomb the heck out of it
-    # Returns
-    # > Boolean, <Nodes> tuple where Boolean says whether there is a clear path to the end and Nodes returns
-    # either the path or the set of closed nodes
-    def calculateAStarPath(self, startNode, endNode, wrld, shouldPathAroundMonsters):
-        openNodes = []
-        closedNodes = []
-        # End node is position 7, 18
-        startNode = Node(startNode.x, startNode.y)
-        openNodes.append(startNode)
-        while len(openNodes) > 0:
-            minNode = openNodes[0]
-            currIndex = 0
-
-            # find the smallest node fval and append it
-            for index, currNode in enumerate(openNodes):
-                if currNode.fval < minNode.fval:
-                    minNode = currNode
-                    currIndex = index
-
-            openNodes.pop(currIndex)
-            closedNodes.append(minNode)
-
-            if minNode.x == endNode.x and minNode.y == endNode.y:
-                path = []
-                currNode = minNode
-                while currNode is not None:
-                    path.append(currNode)
-                    currNode = currNode.parent
-                return True, path[::-1]
-            for neighbor in self.getNeighbors(minNode, endNode, wrld):
-                isClosed = False
-                isOpen = False
-
-                sameNode = None
-                nodeIndex = None
-                for closedNode in closedNodes:
-                    if neighbor.x == closedNode.x and neighbor.y == closedNode.y:
-                        isClosed = True
-                        break
-                for index, openNode in enumerate(openNodes):
-                    if neighbor.x == openNode.x and neighbor.y == openNode.y:
-                        isOpen = True
-                        sameNode = openNode
-                        nodeIndex = index
-                        break
-
-                if not isClosed:
-                    neighbor.gval = minNode.gval + 1
-                    neighbor.hval = self.distanceBetweenNodes(neighbor, endNode, shouldPathAroundMonsters)
-                    neighbor.fval = neighbor.gval + neighbor.hval
-
-                    if isOpen and neighbor.fval < sameNode.fval:
-                        openNodes[nodeIndex] = neighbor
-
-                    elif not isOpen:
-                        openNodes.append(neighbor)
-        return False, closedNodes
-
-    # code taken from selfpreserving_monster.py
-    def getNeighbors(self, node, endNode, wrld):
-        listOfNeighbors = []
-        # Go through neighboring cells
-        for dx in [-1, 0, 1]:
-            # Avoid out-of-bounds access
-            if (node.x + dx >= 0) and (node.x + dx < wrld.width()):
-                for dy in [-1, 0, 1]:
-                    # Avoid out-of-bounds access
-                    if (node.y + dy >= 0) and (node.y + dy < wrld.height()):
-                        # add node if it's the end node, useful if calculating distance to monster
-                        if endNode.x == node.x + dx and endNode.y == node.y + dy:
-                            listOfNeighbors.append(Node(node.x + dx, node.y + dy, parent=node))
-                        # Is this cell not a wall or an explosion?
-                        elif not (wrld.wall_at(node.x + dx, node.y + dy) or
-                                  wrld.explosion_at(node.x + dx, node.y + dy)):
-                            if not (dx == 0 and dy == 0):
-                                listOfNeighbors.append(Node(node.x + dx, node.y + dy, parent=node))
-        # All done
-        return listOfNeighbors
-
-    # absolute distance between nodes
-    def absoluteDistanceBetweenNodes(self, currNode, endNode):
-        xDistance = abs(endNode.x - currNode.x)
-        yDistance = abs(endNode.y - currNode.y)
-
-        return xDistance + yDistance
-
-
-    # distance between two nodes
-    def distanceBetweenNodes(self, currNode, endNode, shouldPathAroundMonsters):
-        xDistance = abs(endNode.x - currNode.x)
-        yDistance = abs(endNode.y - currNode.y)
-
-        # if the node is close to any monster and the endnode is not a monster, try to avoid it
-
-        if shouldPathAroundMonsters:
-            for i in range(0, 4):
-                for monster in self.monsters:
-                    if monster.type == 'stupid' and i < 3:
-                        if (monster.x - i <= currNode.x <= monster.x + i) \
-                                and (monster.y - i <= currNode.y <= monster.y + i):
-                            return max(xDistance, yDistance) + 300 - i * 100
-
-                    else:
-                        if (monster.x - i <= currNode.x <= monster.x + i) \
-                                and (monster.y - i <= currNode.y <= monster.y + i):
-                            return max(xDistance, yDistance) + 300 - i * 100
-
-        # moving diagonally is one move so can combine x and y distance
-        return max(xDistance, yDistance)
